requirements.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
openai
huggingface_hub
InstructorEmbedding
sentence_transformers==2.2.2
chromadb
